[docs] Always use auto-generated apiRef documentation
