+++ docs/pages/api-docs/data-grid/grid-pagination-api.json
@@ -1,0 +1,16 @@
+{
+  "name": "GridPaginationApi",
+  "description": "The pagination API interface that is available in the grid apiRef.",
+  "properties": [
+    {
+      "name": "setPage",
+      "description": "Sets the displayed page to the value given by <code>page</code>.",
+      "type": "(page: number) => void"
+    },
+    {
+      "name": "setPageSize",
+      "description": "Sets the number of displayed rows to the value given by <code>pageSize</code>.",
+      "type": "(pageSize: number) => void"
+    }
+  ]
+}
