+++ docs/scripts/api/buildInterfacesDocumentation.ts
@@ -19,6 +19,8 @@ const INTERFACES_WITH_DEDICATED_PAGES = [
   'GridRowParams',
   'GridSelectionApi',
   'GridFilterApi',
+  'GridSortApi',
+  'GridPaginationApi',
   'GridCsvExportApi',
   'GridCsvExportOptions',
   'GridPrintExportApi',
