+++ docs/src/pages/components/data-grid/sorting/SortingApiNoSnap.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import ApiDocs from 'docsx/src/modules/components/ApiDocs';
+import api from 'docsx/pages/api-docs/data-grid/grid-sort-api.json';
+
+export default function SortingApiNoSnap() {
+  return <ApiDocs api={api} />;
+}
