--- packages/grid/_modules_/grid/hooks/features/pagination/gridPaginationState.ts
@@ -1,6 +1,0 @@
-export interface GridPaginationState {
-  pageSize: number;
-  page: number;
-  pageCount: number;
-  rowCount: number;
-}
