+++ packages/grid/_modules_/grid/hooks/features/pagination/useGridPageSize.ts
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { GridApiRef } from '../../../models';
 import { GridComponentProps } from '../../../GridComponentProps';
-import { GridPageSizeApi } from '../../../models/api/gridPageSizeApi';
+import { GridPageSizeApi } from './gridPaginationInterfaces';
 import { GridEvents } from '../../../constants/eventsConstants';
 import {
   useGridLogger,
