+++ packages/grid/_modules_/grid/models/api/gridApi.ts
@@ -9,8 +9,6 @@ import { GridEditRowApi } from './gridEditRowApi';
 import { GridFilterApi } from './gridFilterApi';
 import { GridFocusApi } from './gridFocusApi';
 import { GridLocaleTextApi } from './gridLocaleTextApi';
-import { GridPageApi } from './gridPageApi';
-import { GridPageSizeApi } from './gridPageSizeApi';
 import { GridParamsApi } from './gridParamsApi';
 import { GridPreferencesPanelApi } from './gridPreferencesPanelApi';
 import { GridPrintExportApi } from './gridPrintExportApi';
@@ -24,6 +22,7 @@ import { GridScrollApi } from './gridScrollApi';
 import type { GridPreProcessingApi } from '../../hooks/core/preProcessing';
 import type { GridRowGroupsPreProcessingApi } from '../../hooks/core/rowGroupsPerProcessing';
 import type { GridDimensionsApi } from '../../hooks/features/dimensions';
+import type { GridPaginationApi } from '../../hooks/features/pagination';
 
 /**
  * The full grid API.
@@ -42,8 +41,7 @@ export interface GridApi
     GridColumnApi,
     GridSelectionApi,
     GridSortApi,
-    GridPageApi,
-    GridPageSizeApi,
+    GridPaginationApi,
     GridCsvExportApi,
     GridFocusApi,
     GridFilterApi,
