+++ scripts/exportsSnapshot.json
@@ -90,8 +90,7 @@
   { "name": "GridLocaleText", "kind": "Interface" },
   { "name": "GridLocaleTextApi", "kind": "Interface" },
   { "name": "GridMenuProps", "kind": "Interface" },
-  { "name": "GridPageApi", "kind": "Interface" },
-  { "name": "GridPageSizeApi", "kind": "Interface" },
+  { "name": "GridPaginationApi", "kind": "Interface" },
   { "name": "GridPaginationState", "kind": "Interface" },
   { "name": "GridPanelClasses", "kind": "Interface" },
   { "name": "GridPanelProps", "kind": "Interface" },
